Rename rule to be prefixed with a11y
