+++ docs/rules/a11y-role-supports-aria-props.md
