+++ lib/rules/a11y-role-supports-aria-props.js
